VideoFaceManager/requirements.txt
# 视频人物识别
opencv-python>=4.8.0
insightface>=0.7.3
onnxruntime>=1.15.0
numpy>=1.24.0
Pillow>=10.0.0

# 数据库
sqlite3-python

# Web面板
Flask>=2.3.0
Werkzeug>=2.3.0

# 工具
click>=8.1.0
tqdm>=4.65.0

# AMD显卡支持（Windows）
# onnxruntime-directml>=1.15.0  # 安装时取消注释
